Improve ripgrep setup to support non-matching
I wrap the :Rg command from fzf.vim. It allows automatically choses the
git root of the currrent file and allows leading arguments of
the form

    :Rg --case-sensitive foobar

to pass arguments to ripgrep. Moreover, quotes like

    :Rg "foo"

are escaped so that ripgrep actually searches for these quotes. This
leads to a problem for

    :Rg ->organisation

because the wrapper interprets this as an argument, and we can't wrap it
in quotes because quotes are escaped into searched-for-quotes. Solution:
Mark that no more arguments follow via --

    :Rg --case-sensitive -- ->organisation